Implement api token authentication

+++ routes/api.php
@@ -15,30 +15,35 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/user', function (Request $request) {
-    return $request->user();
-});
+
+    Route::get('/user', function (Request $request) {
+        return $request->user();
+    });
 
 
-Route::get('/users', function(){
-    return User::all();
-});
+    Route::get('/users', function(){
+        return User::all();
+    });
 
-//Projects
-Route::get('projects', 'ProjectController@index');
-Route::get('projects/{id}', 'ProjectController@show');
-Route::post('projects', 'ProjectController@store');
-Route::put('projects/{id}', 'ProjectController@update');
-Route::delete('projects/{id}', 'ProjectController@destroy');
+    //Projects
+    Route::get('projects', 'ProjectController@index')->middleware('auth:api');
+    Route::get('projects/{id}', 'ProjectController@show')->middleware('auth:api');
+    Route::post('projects', 'ProjectController@store')->middleware('auth:api');
+    Route::put('projects/{id}', 'ProjectController@update')->middleware('auth:api');
+    Route::delete('projects/{id}', 'ProjectController@destroy')->middleware('auth:api');
+
+    //Contractors
+    Route::get('contractors', 'ContractorController@index')->middleware('auth:api');
+    Route::get('contractors/{id}', 'ContractorController@show')->middleware('auth:api');
+    Route::post('contractors', 'ContractorController@store')->middleware('auth:api');
+    Route::put('contractors/{id}', 'ContractorController@update')->middleware('auth:api');
+    Route::delete('contractor/{id}', 'ContractorController@destroy')->middleware('auth:api');
+
+
+
 
 //Administrators
-Route::post('register', 'Auth\RegisterController@register');
-Route::post('login', 'Auth\LoginController@login');
+Route::post('register', 'UsersController@register');
+Route::post('login', 'UsersController@login');
 
-//Contractors
-Route::get('contractors', 'ContractorController@index');
-Route::get('contractors/{id}', 'ContractorController@show');
-Route::post('contractors', 'ContractorController@store');
-Route::put('contractors/{id}', 'ContractorController@update');
-Route::delete('contractor/{id}', 'ContractorController@destroy');
